Scripts/requirements.txt
requests==2.32.4
python-can==4.5.0